refactor: split StellarPasskeyService into focused modules (#976)
* refactor: extract passkey signature helpers into utils module

* refactor: extract passkey transaction helpers into utils module

* refactor: delegate signature and transaction logic to utils modules

* chore: exclude test dir from packages/lib tsconfig rootDir
