+++ packages/lib/package.json
@@ -62,6 +62,9 @@
 	},
 	"main": "./src/index.ts",
 	"types": "./src/index.ts",
+	"scripts": {
+		"test": "bun test"
+	},
 	"license": "MIT",
 	"dependencies": {
 		"@openzeppelin/relayer-plugin-channels": "^0.5.0",
