+++ packages/lib/tsconfig.json
@@ -23,6 +23,6 @@
 		"strict": true,
 		"target": "ES2017"
 	},
-	"exclude": ["node_modules", "dist"],
+	"exclude": ["node_modules", "dist", "test"],
 	"include": ["**/*.ts", "**/*.tsx"]
 }
